+++ src/css/reviews.css
@@ -1,5 +1,5 @@
 /* Стилі для мобільних пристроїв тут */
-.title-text {
+.TitleText {
   font-weight: 600;
   font-size: 16px;
   line-height: 150%;
@@ -10,7 +10,7 @@
   margin-bottom: 32px;
 }
 
-.button-prev {
+.ButtonPrev {
   display: flex;
   padding: 14px;
   border: 1px solid rgba(250, 250, 250, 0.5);
@@ -19,7 +19,7 @@
   stroke: #fafafa;
   transform: rotate(180deg);
 }
-.button-next {
+.ButtonNext {
   display: flex;
   padding: 14px;
   border: 1px solid rgba(250, 250, 250, 0.5);
@@ -40,24 +40,20 @@
   
 }
 /* Картка відгуків */
-.review {
+.Review {
   border: 1px solid rgba(250, 250, 250, 0.2);
   border-radius: 20px;
   padding: 32px 24px;
   flex-basis: calc((100% - 0px) / 1);
-  /* overflow: hidden; */
-  /* height: 302px; */
-  /* flex-shrink: 0; */
-   /* Відключаємо зміщення відгуків, щоб вони не змінювали ширину */
 }
 
 
-.user-icon {
+.UserIcon {
   border-radius: 10px;
   width: 48px;
   height: 48px;
 }
-.name {
+.Name {
   font-weight: 600;
   font-size: 20px;
   line-height: 150%;
@@ -66,29 +62,29 @@
   padding-bottom: 8px;
   padding-top: 32px;
 }
-.review-text {
+.ReviewText {
   font-weight: 500;
   font-size: 16px;
   line-height: 150%;
-  height: 158px;
+  height: 188px;
   letter-spacing: -0.02em;
   color: rgba(250, 250, 250, 0.6);
 }
 /* Картка відгуків ================*/
 
 @media screen and (max-width: 767px) {
-  .arrow-nav {
+  .ArrowNav {
     display: flex;
     flex-direction: row;
     justify-content: center;
     align-items: center;
     gap: 12px;
   }
-  .section {
+  .Section {
     padding-top: 32px;
     padding-bottom: 32px;
   }
-  .container {
+  .Container {
     max-width: 375px;
     margin: 0 auto;
     padding: 0 15px;
@@ -97,18 +93,18 @@
 
 /* Планшети (з 768px) */
 @media screen and (min-width: 768px) {
-  .arrow-nav {
+  .ArrowNav {
     display: flex;
     flex-direction: row;
     justify-content: center;
     align-items: center;
     gap: 12px;
   }
-  .section {
+  .Section {
     padding-top: 48px;
     padding-bottom: 48px;
   }
-  .container {
+  .Container {
     max-width: 768px;
     margin: 0 auto;
     padding: 0 15px;
@@ -118,33 +114,32 @@
      gap: 16px;
   
     }
-    .review {
+    .Review {
       border: 1px solid rgba(250, 250, 250, 0.2);
       border-radius: 20px;
       padding: 32px 24px;
-      /* flex-basis: calc((100% - 16px) / 2); */
       flex-basis: 360px;
-      height: 302px;
-      flex-shrink: 0;
+      display: flex;
+        flex-direction: column;
       /* Відключаємо зміщення відгуків, щоб вони не змінювали ширину */
     }
 
 }
 
-/* Десктоп (з 1280px) */
-@media screen and (min-width: 1280px) {
-  .arrow-nav {
+/* Десктоп (з 1440px) */
+@media screen and (min-width: 1440px) {
+  .ArrowNav {
     display: flex;
     flex-direction: row;
     justify-content: center;
     align-items: center;
     gap: 12px;
   }
-  .section {
+  .Section {
     padding-top: 82px;
     padding-bottom: 82px;
   }
-  .container {
+  .Container {
     max-width: 1280px;
     margin: 0 auto;
     padding: 0 15px;
@@ -155,10 +150,14 @@
   
     }
 
-  .review {
+  .Review {
     display: flex;
     flex-direction: column;
     flex-basis: calc((100% - 48px) / 4);
+}
+.ButtonNext:hover,
+.ButtonPrev:hover {
+  stroke: rgb(59, 59, 59);
 }
     
 }
@@ -171,6 +170,6 @@
   border-color: rgba(250, 250, 250, 0.2);
 }
 
-.social-icon-next {
+.SocialIconBack, .SocialIconNext {
   stroke: inherit;
 }
