+++ src/js/api.js
@@ -22,3 +22,23 @@ export async function PostRequest(email, comment) {
     return resp.json();
   });
 }
+
+export function doFetch(foo, murkUpFoo) {
+const REVIEWS_LIST = document.getElementById('reviewsList');
+fetch("https://portfolio-js.b.goit.study/api/reviews")
+  .then((response) => {
+    if (!response.ok) {
+      throw new Error(response.status);
+    }
+    return response.json();
+  })
+  .then((reviews) => {
+    const markup = reviews.map((review) => {
+      return murkUpFoo(review)
+    }).join("");
+
+    REVIEWS_LIST.innerHTML = markup; // Вставляємо HTML у список відгуків
+    foo()
+  })
+  .catch((error) => console.log(error));
+}
