+++ src/js/reviews.js
@@ -1,71 +1,55 @@
 // import Swiper bundle with all modules installed
 import Swiper from 'swiper';
+import { doFetch } from './api';
 
 // import styles bundle
 import 'swiper/css';
 
-
-const reviewsList = document.getElementById('reviewsList');
-const screenWidth = window.innerWidth;
+const SCREEN_WIDTH = window.innerWidth;
 let slidesPerView;
 
-if (screenWidth >= 1280) {
+if (SCREEN_WIDTH >= 1440) {
   slidesPerView = 4;
-} else if (screenWidth >= 768 && screenWidth < 1280) {
+} else if (SCREEN_WIDTH >= 768 && SCREEN_WIDTH < 1440) {
   slidesPerView = 2;
-} else if (screenWidth < 768) {
+} else if (SCREEN_WIDTH < 768) {
   slidesPerView = 1;
 }
 
 function createSwiper() {
-console.log(screenWidth)
-const swiper = new Swiper('.reviews-swiper', {
+const SWIPER = new Swiper('.reviews-swiper', {
   // Optional parameters
   cssMode: true, 
   slidesPerView: slidesPerView,
   slidesPerGroup: 1,
-  
-
   // Navigation arrows
   navigation: {
     disabledClass: "BtnOff",
-    nextEl: '.button-next',
-    prevEl: '.button-prev',
+    nextEl: '.ButtonNext',
+    prevEl: '.ButtonPrev',
   },
   mousewheel: true,
   keyboard: true,
   touch: true,
-  
-});
+  });
+}
 
-
+function doMurkUp(review) {
+  return`
+        <div class="Review swiper-slide">
+        <img class="UserIcon" src="${review.avatar_url}" alt="Avatar">
+          <h4 class="Name">${review.author}</h4>
+          <p class="ReviewText">${review.review}</p>
+        </div>
+      `
 }
 
 // запит з бекенду=========================================================
-// const reviewsList = document.getElementById('reviewsList');
 
-fetch("https://portfolio-js.b.goit.study/api/reviews")
-  .then((response) => {
-    if (!response.ok) {
-      throw new Error(response.status);
-    }
-    return response.json();
-  })
-  .then((reviews) => {
-    const markup = reviews.map((review) => {
-      return `
-        <div class="review swiper-slide">
-        <img class="user-icon" src="${review.avatar_url}" alt="Avatar">
-          <h4 class="name">${review.author}</h4>
-          <p class="review-text">${review.review}</p>
-        </div>
-      `;
-    }).join("");
+doFetch(createSwiper, doMurkUp)
 
-    reviewsList.innerHTML = markup; // Вставляємо HTML у список відгуків
-    createSwiper()
-  })
-  .catch((error) => console.log(error));
+
+
