+++ src/partials/reviews.html
@@ -1,21 +1,21 @@
-<section class="section section-reviews">
-  <div class="container">
-    <h3 class="title-text">REVIEWS</h3>
+<section class="Section SectionReviews">
+  <div class="Container">
+    <h3 class="TitleText">REVIEWS</h3>
 
     <!-- Swiper -->
     <div class="reviews-swiper">
       <div id="reviewsList" class="swiper-wrapper">
         <!-- тут додані відгуки -->
       </div>
-      <div class="arrow-nav">
-        <div class="button-prev">
-          <svg class="social-icon-back" width="32" height="32">
+      <div class="ArrowNav">
+        <div class="ButtonPrev">
+          <svg class="SocialIconBack" width="32" height="32">
             <use href="./img/icons.svg#icon-arrow-narrow-right"></use>
           </svg>
         </div>
 
-        <div class="button-next">
-          <svg class="social-icon-next" width="32" height="32">
+        <div class="ButtonNext">
+          <svg class="SocialIconNext" width="32" height="32">
             <use href="./img/icons.svg#icon-arrow-narrow-right"></use>
           </svg>
         </div>
